toxic_detect_api/requirements.txt
djangorestframework
markdown
django-filter
tensorflow
contractions
nltk
djangorestframework-simplejwt
django-cors-headers
django-environ